+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 - Changed: dropped support for php lower than 8.1
 - Removed: support for deprecated config key videoProvider
 - Removed: support for deprecated config option names
+- Removed: support for deprecated privacy mode attribute
+- Removed: support for deprecated huh.video.event.alertify.onshow and huh.video.event.alertify.onfocus events
 
 ### Upgrade steps
 - If you have custom fullsize or privacy templates, move and rename them according to the new contao template system